requirements.txt
Flask==1.0.2
Flask-WTF==0.14.2
PyMySQL==0.9.3
Flask-SQLAlchemy==2.3.2
WTForms==2.2.1